+++ app/setup/component/componentOperatorInstallerStep_test.go
@@ -36,14 +36,15 @@ func TestComponentOperatorInstallerStep_GetStepDescription(t *testing.T) {
 	t.Run("should get description", func(t *testing.T) {
 		// given
 		step := &componentOperatorInstallerStep{
-			chart: "testChart",
+			chart:    "testChart",
+			crdChart: "testCrdChart",
 		}
 
 		// when
 		desc := step.GetStepDescription()
 
 		// then
-		assert.Equal(t, "Install component-operator from testChart", desc)
+		assert.Equal(t, "Install component-operator from testChart and component-crd from testCrdChart", desc)
 	})
 }
 
@@ -51,6 +52,15 @@ func TestComponentOperatorInstallerStep_PerformSetupStep(t *testing.T) {
 	t.Run("should successfully perform setup", func(t *testing.T) {
 		// given
 		testCtx := context.TODO()
+
+		crdChartSpec := &helmclient.ChartSpec{
+			ReleaseName: "testCrdChart",
+			ChartName:   "foo/testCrdChart",
+			Namespace:   "testing",
+			Version:     "0.1",
+			Timeout:     time.Second * 300,
+			Wait:        true,
+		}
 
 		chartSpec := &helmclient.ChartSpec{
 			ReleaseName: "testChart",
@@ -62,11 +72,13 @@ func TestComponentOperatorInstallerStep_PerformSetupStep(t *testing.T) {
 		}
 
 		helmClientMock := newMockHelmClient(t)
+		helmClientMock.EXPECT().InstallOrUpgrade(testCtx, crdChartSpec).Return(nil)
 		helmClientMock.EXPECT().InstallOrUpgrade(testCtx, chartSpec).Return(nil)
 
 		step := &componentOperatorInstallerStep{
 			namespace:  "testing",
 			chart:      "foo/testChart:0.1",
+			crdChart:   "foo/testCrdChart:0.1",
 			helmClient: helmClientMock,
 		}
 
@@ -75,10 +87,20 @@ func TestComponentOperatorInstallerStep_PerformSetupStep(t *testing.T) {
 
 		// then
 		require.NoError(t, err)
+		helmClientMock.AssertNumberOfCalls(t, "InstallOrUpgrade", 2)
 	})
 	t.Run("should successfully perform setup for 'latest' version", func(t *testing.T) {
 		// given
 		testCtx := context.TODO()
+
+		crdChartSpec := &helmclient.ChartSpec{
+			ReleaseName: "testCrdChart",
+			ChartName:   "foo/testCrdChart",
+			Namespace:   "testing",
+			Version:     "",
+			Timeout:     time.Second * 300,
+			Wait:        true,
+		}
 
 		chartSpec := &helmclient.ChartSpec{
 			ReleaseName: "testChart",
@@ -90,11 +112,13 @@ func TestComponentOperatorInstallerStep_PerformSetupStep(t *testing.T) {
 		}
 
 		helmClientMock := newMockHelmClient(t)
+		helmClientMock.EXPECT().InstallOrUpgrade(testCtx, crdChartSpec).Return(nil)
 		helmClientMock.EXPECT().InstallOrUpgrade(testCtx, chartSpec).Return(nil)
 
 		step := &componentOperatorInstallerStep{
 			namespace:  "testing",
 			chart:      "foo/testChart:latest",
+			crdChart:   "foo/testCrdChart:latest",
 			helmClient: helmClientMock,
 		}
 
@@ -103,8 +127,9 @@ func TestComponentOperatorInstallerStep_PerformSetupStep(t *testing.T) {
 
 		// then
 		require.NoError(t, err)
+		helmClientMock.AssertNumberOfCalls(t, "InstallOrUpgrade", 2)
 	})
-	t.Run("should fail to perform setup for wrong chart format", func(t *testing.T) {
+	t.Run("should fail to perform setup for wrong crd chart format", func(t *testing.T) {
 		// given
 		testCtx := context.TODO()
 
@@ -112,7 +137,7 @@ func TestComponentOperatorInstallerStep_PerformSetupStep(t *testing.T) {
 
 		step := &componentOperatorInstallerStep{
 			namespace:  "testing",
-			chart:      "foo/testChart_0.1",
+			crdChart:   "foo/testCrdChart_0.1",
 			helmClient: helmClientMock,
 		}
 
@@ -121,9 +146,9 @@ func TestComponentOperatorInstallerStep_PerformSetupStep(t *testing.T) {
 
 		// then
 		require.Error(t, err)
-		require.ErrorContains(t, err, "componentOperatorChart 'foo/testChart_0.1' has a wrong format. Must be '<chartName>:<version>'; e.g.: 'foo/bar:1.2.3'")
+		require.ErrorContains(t, err, "componentChart 'foo/testCrdChart_0.1' has a wrong format. Must be '<chartName>:<version>'; e.g.: 'foo/bar:1.2.3'")
 	})
-	t.Run("should fail to perform setup for empty chartName", func(t *testing.T) {
+	t.Run("should fail to perform setup for empty crd chartName", func(t *testing.T) {
 		// given
 		testCtx := context.TODO()
 
@@ -131,7 +156,7 @@ func TestComponentOperatorInstallerStep_PerformSetupStep(t *testing.T) {
 
 		step := &componentOperatorInstallerStep{
 			namespace:  "testing",
-			chart:      "foo/:0.1",
+			crdChart:   "foo/:0.1",
 			helmClient: helmClientMock,
 		}
 
@@ -142,12 +167,12 @@ func TestComponentOperatorInstallerStep_PerformSetupStep(t *testing.T) {
 		require.Error(t, err)
 		require.ErrorContains(t, err, "error reading chartname 'foo/': wrong format")
 	})
-	t.Run("should fail to perform setup for error in helmClient", func(t *testing.T) {
+	t.Run("should fail to perform setup for error in helmClient when installing the crd chart", func(t *testing.T) {
 		// given
 		testCtx := context.TODO()
 		chartSpec := &helmclient.ChartSpec{
-			ReleaseName: "testChart",
-			ChartName:   "foo/testChart",
+			ReleaseName: "testCrdChart",
+			ChartName:   "foo/testCrdChart",
 			Namespace:   "testing",
 			Version:     "0.1",
 			Timeout:     time.Second * 300,
@@ -159,7 +184,106 @@ func TestComponentOperatorInstallerStep_PerformSetupStep(t *testing.T) {
 
 		step := &componentOperatorInstallerStep{
 			namespace:  "testing",
+			crdChart:   "foo/testCrdChart:0.1",
+			helmClient: helmClientMock,
+		}
+
+		// when
+		err := step.PerformSetupStep(testCtx)
+
+		// then
+		require.Error(t, err)
+		require.ErrorIs(t, err, assert.AnError)
+	})
+	t.Run("should fail to perform setup for wrong operator chart format", func(t *testing.T) {
+		// given
+		testCtx := context.TODO()
+
+		crdChartSpec := &helmclient.ChartSpec{
+			ReleaseName: "testCrdChart",
+			ChartName:   "foo/testCrdChart",
+			Namespace:   "testing",
+			Version:     "0.1",
+			Timeout:     time.Second * 300,
+			Wait:        true,
+		}
+
+		helmClientMock := newMockHelmClient(t)
+		helmClientMock.EXPECT().InstallOrUpgrade(testCtx, crdChartSpec).Return(nil)
+
+		step := &componentOperatorInstallerStep{
+			namespace:  "testing",
+			chart:      "foo/testChart_0.1",
+			crdChart:   "foo/testCrdChart:0.1",
+			helmClient: helmClientMock,
+		}
+
+		// when
+		err := step.PerformSetupStep(testCtx)
+
+		// then
+		require.Error(t, err)
+		require.ErrorContains(t, err, "componentChart 'foo/testChart_0.1' has a wrong format. Must be '<chartName>:<version>'; e.g.: 'foo/bar:1.2.3'")
+	})
+	t.Run("should fail to perform setup for empty operator chartName", func(t *testing.T) {
+		// given
+		testCtx := context.TODO()
+
+		crdChartSpec := &helmclient.ChartSpec{
+			ReleaseName: "testCrdChart",
+			ChartName:   "foo/testCrdChart",
+			Namespace:   "testing",
+			Version:     "0.1",
+			Timeout:     time.Second * 300,
+			Wait:        true,
+		}
+
+		helmClientMock := newMockHelmClient(t)
+		helmClientMock.EXPECT().InstallOrUpgrade(testCtx, crdChartSpec).Return(nil)
+
+		step := &componentOperatorInstallerStep{
+			namespace:  "testing",
+			chart:      "foo/:0.1",
+			crdChart:   "foo/testCrdChart:0.1",
+			helmClient: helmClientMock,
+		}
+
+		// when
+		err := step.PerformSetupStep(testCtx)
+
+		// then
+		require.Error(t, err)
+		require.ErrorContains(t, err, "error reading chartname 'foo/': wrong format")
+	})
+	t.Run("should fail to perform setup for error in helmClient when installing the operator chart", func(t *testing.T) {
+		// given
+		testCtx := context.TODO()
+		chartSpec := &helmclient.ChartSpec{
+			ReleaseName: "testChart",
+			ChartName:   "foo/testChart",
+			Namespace:   "testing",
+			Version:     "0.1",
+			Timeout:     time.Second * 300,
+			Wait:        true,
+		}
+
+		crdChartSpec := &helmclient.ChartSpec{
+			ReleaseName: "testCrdChart",
+			ChartName:   "foo/testCrdChart",
+			Namespace:   "testing",
+			Version:     "0.1",
+			Timeout:     time.Second * 300,
+			Wait:        true,
+		}
+
+		helmClientMock := newMockHelmClient(t)
+		helmClientMock.EXPECT().InstallOrUpgrade(testCtx, crdChartSpec).Return(nil)
+		helmClientMock.EXPECT().InstallOrUpgrade(testCtx, chartSpec).Return(assert.AnError)
+
+		step := &componentOperatorInstallerStep{
+			namespace:  "testing",
 			chart:      "foo/testChart:0.1",
+			crdChart:   "foo/testCrdChart:0.1",
 			helmClient: helmClientMock,
 		}